+++ enlighten/measurement/AreaScanFeature.py
@@ -1,11 +1,11 @@
 import os
 import logging
-import datetime
 import pyqtgraph
 import numpy as np
 import qimage2ndarray
 
 from PIL import Image, ImageStat
+from datetime import datetime
 
 from enlighten import common
 from enlighten.ui.ScrollStealFilter import ScrollStealFilter
@@ -53,6 +53,14 @@ class AreaScanFeature:
     include each full-frame image along with the vertically binned spectrum,
     so the image can be directly displayed in ENLIGHTEN.
 
+    @par Batch Collection
+
+    I am not taking time to fully integrate Area Scan into our BatchCollection
+    at this time -- I think there are multiple reasons why that might prove 
+    tricky, and it's too much to take on right now for the limited target
+    audience. Instead I'm kludging in a very lightweight connection between
+    the two.
+
     """
 
     # ##########################################################################
@@ -66,6 +74,7 @@ class AreaScanFeature:
         self.bt_save            = cfu.pushButton_area_scan_save
         self.cb_enable          = cfu.checkBox_area_scan_enable
         self.cb_normalize       = cfu.checkBox_area_scan_normalize
+        self.cb_normalize_csv   = cfu.checkBox_area_scan_normalize_csv
         self.cb_fit             = cfu.checkBox_area_scan_fit
         self.frame_image        = cfu.frame_area_scan_image
         self.frame_live         = cfu.frame_area_scan_live
@@ -78,14 +87,20 @@ class AreaScanFeature:
         self.sb_start           = cfu.spinBox_area_scan_start_line
         self.sb_stop            = cfu.spinBox_area_scan_stop_line
         self.sb_step            = cfu.spinBox_area_scan_line_step
+        self.sb_scale           = cfu.spinBox_area_scan_scale
+        self.sb_cursor          = cfu.spinBox_area_scan_cursor
 
         self.data = None
+        self.data_raw = None
         self.enabled = False
         self.visible = False
         self.normalize = True
+        self.normalize_csv = True
         self.fit = True
         self.start_line = 0
         self.stop_line = 63
+        self.scale = 1
+        self.cursor_line = 0
         self.ignored = 0
         self.frame_count = 0
         self.image = None
@@ -94,27 +109,35 @@ class AreaScanFeature:
         self.last_line = 0
         self.last_elapsed_sec = 0
         self.curve_live = None
-        self.frame_start = datetime.datetime.now()
+        self.frame_start = datetime.now()
+        self.last_save_timestamp = self.frame_start
 
         self.pen_start  = self.ctl.gui.make_pen(color="enlighten_name_g",  width=2)
         self.pen_stop   = self.ctl.gui.make_pen(color="enlighten_name_n2", width=2)
         self.pen_line   = self.ctl.gui.make_pen(color="enlighten_name_h")
-        self.pen_cursor = self.ctl.gui.make_pen(color="enlighten_name_e1")
+        self.pen_cursor = self.ctl.gui.make_pen(color="enlighten_name_i")
+
+        self.cb_fit.setEnabled(True)
+        self.sb_scale.setEnabled(True)
 
         # create widgets we can't / don't pass in
         self.create_widgets()
 
         self.cb_fit.setChecked(True)
         self.cb_normalize.setChecked(True)
+        self.cb_normalize_csv.setChecked(True)
         self.progress_bar.setVisible(False)
 
-        self.bt_save     .clicked        .connect(self.save_callback)
+        self.bt_save     .clicked        .connect(self.save)
         self.cb_normalize.stateChanged   .connect(self.normalize_callback)
+        self.cb_normalize_csv.stateChanged.connect(self.normalize_callback)
         self.cb_fit      .stateChanged   .connect(self.fit_callback)
         self.cb_enable   .stateChanged   .connect(self.enable_callback)
         self.sb_start    .valueChanged   .connect(self.roi_callback)
         self.sb_stop     .valueChanged   .connect(self.roi_callback)
         self.sb_step     .valueChanged   .connect(self.step_callback)
+        self.sb_scale    .valueChanged   .connect(self.scale_callback)
+        self.sb_cursor   .valueChanged   .connect(self.cursor_callback)
 
         self.progress_bar_timer = QtCore.QTimer()
         self.progress_bar_timer.timeout.connect(self.tick_progress_bar)
@@ -208,6 +231,14 @@ class AreaScanFeature:
     # callbacks
     # ##########################################################################
 
+    def cursor_callback(self):
+        self.cursor_line = self.sb_cursor.value()
+
+    def scale_callback(self):
+        self.scale = self.sb_scale.value()
+        self.cb_fit.setChecked(False)
+        self.cb_fit.setEnabled(False)
+
     def step_callback(self):
         """ the user changed the step spinner """
         spec = self.ctl.multispec.current_spectrometer()
@@ -231,6 +262,7 @@ class AreaScanFeature:
         if spec is None:
             return self.disable()
         self.normalize = self.cb_normalize.isChecked()
+        self.normalize_csv = self.cb_normalize_csv.isChecked()
 
     def fit_callback(self):
         """ The user clicked "[x] fit" on the widget """
@@ -243,6 +275,7 @@ class AreaScanFeature:
         log.debug("disabling area scan")
         self.enabled = False
         self.data = None
+        self.data_raw = None
         self.last_received_time = None
         self.frame_image.setVisible(False)
         self.cb_enable.setChecked(False)
@@ -274,40 +307,56 @@ class AreaScanFeature:
             self.update_from_gui()
             self.ignored = 0
 
-    def save_callback(self):
+    def save(self):
         spec = self.ctl.multispec.current_spectrometer()
         if spec is None:
             return self.disable()
 
+        saved_something = False
         today_dir = self.ctl.save_options.generate_today_dir()
-        basename = "area-scan-%s-%s" % (datetime.datetime.now().strftime("%Y%m%d-%H%M%S"),
-                                        spec.settings.eeprom.serial_number)
+
+        # this doesn't use the full templating capability of Measurement (which 
+        # needs extracted into TemplateFeature), but meets the immediate need
+        now = datetime.now()
+        ts = now.strftime("%Y%m%d-%H%M%S")
+        sn = spec.settings.eeprom.serial_number
+        basename = f"area-scan-{ts}-{sn}"
+        if self.ctl.save_options.has_prefix():
+            basename = self.ctl.save_options.prefix() + "-" + basename
+        if self.ctl.save_options.has_suffix():
+            basename += "-" + self.ctl.save_options.suffix()
 
         # save image
         if self.image is not None:
             pathname_png = os.path.join(today_dir, basename + ".png")
             log.debug("saving qimage %s", pathname_png)
             self.image.save(pathname_png)
+            saved_something = True
 
-        # save table
-        pathname_csv = os.path.join(today_dir, basename + ".csv")
-        log.debug("saving csv %s", pathname_csv)
+        data = self.data if self.normalize_csv else self.data_raw
+        if data is not None:
+            # save table
+            pathname_csv = os.path.join(today_dir, basename + ".csv")
+            log.debug("saving csv %s", pathname_csv)
 
-        lines = len(self.data)
-        pixels = len(self.data[0])
-        with open(pathname_csv, "w") as outfile:
-            for i in range(pixels):
-                outfile.write(f", {i}")
-            outfile.write("\n")
-            for line in range(lines):
-                outfile.write(f"{line}")
+            lines = len(data)
+            pixels = len(data[0])
+            with open(pathname_csv, "w") as outfile:
                 for i in range(pixels):
-                    outfile.write(f", {self.data[line][i]}")
+                    outfile.write(f", {i}")
                 outfile.write("\n")
+                for line in range(lines):
+                    outfile.write(f"{line}")
+                    for i in range(pixels):
+                        outfile.write(f", {data[line][i]}")
+                    outfile.write("\n")
+            saved_something = True
 
-        # np.savetxt(pathname_csv, self.data, fmt="%d", delimiter=",")
-
-        self.ctl.marquee.info("saved %s" % basename)
+        if saved_something:
+            self.ctl.marquee.info("saved %s" % basename)
+            self.last_save_timestamp = now
+        else:
+            self.ctl.marquee.error("no area scan data to save")
 
     # ##########################################################################
     # private methods
@@ -334,14 +383,14 @@ class AreaScanFeature:
         """ we've updated the start/stop lines, so resize the image """
         if self.last_received_time is None:
             log.debug("update_progress_bar: initializing")
-            self.last_received_time = datetime.datetime.now()
+            self.last_received_time = datetime.now()
             return
 
-        elapsed_ms = round(1000 * (datetime.datetime.now() - self.last_received_time).total_seconds())
+        elapsed_ms = round(1000 * (datetime.now() - self.last_received_time).total_seconds())
         log.debug("update_progress_bar: elapsed = %d ms (estimate was %d)", elapsed_ms, self.progress_bar.maximum())
         if elapsed_ms > 5000:
             log.debug("update_progress_bar: ignoring overlong elapsed")
-            self.last_received_time = datetime.datetime.now()
+            self.last_received_time = datetime.now()
             return
     
         self.progress_bar.setVisible(True)
@@ -350,7 +399,7 @@ class AreaScanFeature:
         self.progress_bar.setValue(1)
         self.progress_bar_timer.start(100)
 
-        self.last_received_time = datetime.datetime.now()
+        self.last_received_time = datetime.now()
 
     def tick_progress_bar(self):
         if not self.enabled:
@@ -360,7 +409,7 @@ class AreaScanFeature:
             return
         elif self.last_received_time is None:
             return
-        elapsed_ms = round(1000 * (datetime.datetime.now() - self.last_received_time).total_seconds())
+        elapsed_ms = round(1000 * (datetime.now() - self.last_received_time).total_seconds())
         self.progress_bar.setValue(elapsed_ms)
         self.progress_bar_timer.start(100)
 
@@ -413,6 +462,7 @@ class AreaScanFeature:
             h, w = data.shape
 
             # optionally normalize
+            self.data_raw = data.astype(np.uint16)
             if self.normalize:
                 lo = np.min(data)
                 hi = np.max(data)
@@ -425,6 +475,15 @@ class AreaScanFeature:
             self.image = QtGui.QImage(data, w, h, w*2, QtGui.QImage.Format_Grayscale16)
             qpixmap = QtGui.QPixmap.fromImage(self.image)
 
+            # scale image if requested
+            # - QPixmap.setDevicePixelRatio()
+            # - QPixmap.scaled()
+            # - QImage.scaled()
+            if self.scale > 1:
+                # note that this doesn't change self.image, which is what gets saved as a PNG
+                size = QtCore.QSize(w * self.scale, h * self.scale)
+                qpixmap = qpixmap.scaled(size)
+
             self.scene.clear()
             self.scene.addPixmap(qpixmap)
 
@@ -434,11 +493,19 @@ class AreaScanFeature:
             # add three horizontal marker lines ATOP the image to show bounds and progress
             log.debug(f"adding green/red lines at (start {start}, stop {stop}) on image (width {w}, height {h}), line_index {line_index}")
             margin = 15
-            self.scene.addLine(-margin, start,      w + margin, start,      self.pen_start)
-            self.scene.addLine(-margin, line_index, w + margin, line_index, self.pen_line) 
-            self.scene.addLine(-margin, stop,       w + margin, stop,       self.pen_stop)
 
-            self.display_cursor(y0=-margin, y1=h+margin)
+            scaled_w     = self.scale * w
+            scaled_h     = self.scale * h
+            scaled_start = self.scale * start
+            scaled_index = self.scale * line_index
+            scaled_stop  = self.scale * stop
+
+            self.scene.addLine(-margin, scaled_start, scaled_w + margin, scaled_start, self.pen_start)
+            self.scene.addLine(-margin, scaled_index, scaled_w + margin, scaled_index, self.pen_line) 
+            self.scene.addLine(-margin, scaled_stop,  scaled_w + margin, scaled_stop,  self.pen_stop)
+
+            self.display_vertical_cursor  (y0 = -margin, y1 = margin + scaled_h)
+            self.display_horizontal_cursor(x0 = -margin, x1 = margin + scaled_w)
 
             # optionally scale entire QGraphicsView
             t = QtGui.QTransform()
@@ -455,9 +522,9 @@ class AreaScanFeature:
 
             # display timing
             if line_index < self.last_line:
-                self.last_elapsed_sec = (datetime.datetime.now() - self.frame_start).total_seconds()
-                self.frame_start = datetime.datetime.now()
-            elapsed_sec = (datetime.datetime.now() - self.frame_start).total_seconds()
+                self.last_elapsed_sec = (datetime.now() - self.frame_start).total_seconds()
+                self.frame_start = datetime.now()
+            elapsed_sec = (datetime.now() - self.frame_start).total_seconds()
             self.lb_elapsed.setText(f"{elapsed_sec:.2f} (last {self.last_elapsed_sec:.2f})")
 
             self.last_line = line_index
@@ -465,6 +532,38 @@ class AreaScanFeature:
 
         except:
             log.error("failed to render AreaScanImage data", exc_info=1)
+
+        self.check_for_batch_collection()
+
+    def check_for_batch_collection(self):
+        """
+        This is as far as we've currently integrated Area Scan into Batch 
+        Collection. Basically, we use the timing periods and counts from the 
+        BatchCollection form when BatchCollection is enabled, but do not attempt
+        to ride over the whole "TakeOneRequest" / VCRControls pipeline. I'm sure
+        it's doable, but I'm not doing it now.
+        """
+        if not self.enabled:
+            # we only auto-save Area Scan if the feature is enabled
+            return
+
+        if not self.ctl.page_nav.doing_factory():
+            # we only auto-save Area Scan when we're looking at area scan
+            return
+
+        bc = self.ctl.batch_collection
+        if not bc.enabled:
+            # we only auto-save Area Scan if BatchCollection is ENABLED 
+            # (which is not the same as RUNNING)
+            return
+
+        period_ms = bc.measurement_period_ms
+        elapsed_ms = (datetime.now() - self.last_save_timestamp).total_seconds() * 1000.0
+        if period_ms > elapsed_ms:
+            # use BatchCollection "measurement period" to decide how often to save an Area Scan
+            return
+
+        self.save()
 
     def process_reading_with_area_scan_image_png(self, reading):
         """ we have received a Reading with an AreaScanImage with pathname_png populated """
@@ -484,20 +583,28 @@ class AreaScanFeature:
             if asi.height is not None and asi.height_orig is not None:
                 scale = 1.0 * asi.height / asi.height_orig
                 
+            margin = 10
             start_line = spec.settings.eeprom.roi_vertical_region_1_start
             stop_line = spec.settings.eeprom.roi_vertical_region_1_end
 
-            self.scene.addLine(-10, scale*start_line, qpixmap.width() + 10, scale*start_line, self.pen_start)
-            self.scene.addLine(-10, scale*stop_line,  qpixmap.width() + 10, scale*stop_line,  self.pen_stop)
+            self.scene.addLine(-margin, scale*start_line, qpixmap.width() + margin, scale*start_line, self.pen_start)
+            self.scene.addLine(-margin, scale*stop_line,  qpixmap.width() + margin, scale*stop_line,  self.pen_stop)
 
-            self.display_cursor(y0=-10, y1=qpixmap.height() + 10)
+            self.display_vertical_cursor  (y0 = -margin, y1 = margin + qpixmap.height())
+            self.display_horizontal_cursor(x0 = -margin, x1 = margin + qpixmap.width())
         except:
             log.error("failed to display PNG", exc_info=1)
 
-    def display_cursor(self, y0, y1):
+    def display_vertical_cursor(self, y0, y1):
         if self.ctl.cursor.is_enabled():
             x = self.ctl.cursor.get_pixel()
             self.scene.addLine(x, y0, x, y1, self.pen_cursor)
+
+    def display_horizontal_cursor(self, x0, x1):
+        y = self.cursor_line * self.scale
+        # log.debug(f"display_horizontal_cursor: y {y}, x0 {x0}, x1 {x1}")
+        if y > 0:
+            self.scene.addLine(x0, y, x1, y, self.pen_cursor)
 
     def normalize_png(self, pathname_png):
         try:
@@ -533,12 +640,9 @@ class AreaScanFeature:
 
         log.debug("resize: data_w = %d, data_h = %d (start %d, stop %d)", data_w, data_h, self.start_line, self.stop_line)
         self.data = np.zeros((data_h, data_w), dtype=np.float32)
+        self.data_raw = None
 
-        self.frame_image.setMinimumHeight(data_h + 40)
         self.graphics_view.setMinimumHeight(150)
-
-        # if spec.settings.is_micro():
-        #    self.image.move(0, 0)
 
     def finish_update(self):
         """
@@ -555,6 +659,6 @@ class AreaScanFeature:
         self.scene.clear() # @todo - anything leak here? need to deleteLater old pixmap?
         self.scene.addPixmap(pixmap)
 
-        # vertically bin the on-screen image for the "live" spectrum
+        # vertically bin the on-screen image for the "live" spectrum (optionally normalized)
         total = np.sum(self.data, axis=0)
         self.ctl.set_curve_data(self.curve_live, total)